accounts_api/requirements.txt
Django==3.1.1
django-querycount==0.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1

